Update patch set 2
Patch Set 2:

Ah, so it will require forking in a new repo.

Adding pulser for this.  We need to make sure that the IR app doesn't misbehave on devices without IR (yuga).  I'm going to keep my -1 for now as a reminder this patch could be a problem on yuga for testing/checking - plus this commit should be squashed into the one that adds the original name to PRODUCT_PACKAGEs and that commit's commit message should indicate the source of the IRRemote package

Patch-set: 2