Update patch set 1
Patch Set 1:

Without this commit, the device will boot normally after a charger has been connected. However, it seems like it "thinks" it's in some offboard charging mode and turns off as soon as the charger gets disconnected.
With this commit, the bootanimation is shown for 1-2 seconds (which is annoying and the reason for my "-1"), but the device does not actually boot, the screen stays black with the LED turned on.

Patch-set: 1