fusion3: policy updates
Fixes radio start on boot with empty /cache/qcks, as well as some
system_monitor denials and a phone app crash on calls related
to the wake_gesture sysfs node. Also lets mediaserver read the
ALS, to make it happy and it lets the settings app modify the
vibrator levels.

Change-Id: Idade0a21e579e292f4da13ec217e45d1e0adea85